# video-editor/storage/api-docs/api-docs.json
{
  "swagger":"2.0",
  "info":{
    "description":"",
    "version":"1.0.0",
    "title":"Swagger Video Editor API",
    "termsOfService":"http://swagger.io/terms/",
    "contact":{
      "email":"apiteam@swagger.io"
    },
    "license":{
      "name":"Apache 2.0",
      "url":"http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host":"app.loc:8080",
  "basePath":"/api",
  "schemes":[
    "http",
    "https"
  ],
  "paths":{
    "/video":{
      "post":{
        "tags":[
          "video"
        ],
        "summary":"Create a new video",
        "description":"",
        "operationId":"createVideo",
        "consumes":[
          "application/json",
          "application/xml"
        ],
        "produces":[
          "application/xml",
          "application/json"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"body",
            "description":"Video object that needs to be added to the store",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Video"
            }
          }
        ],
        "responses":{
          "405":{
            "description":"Invalid input"
          }
        }
      }
    }
  },
  "definitions": {
    "Video": {
      "type": "object",
      "required": [
        "first_url",
        "first_start",
        "first_duration",
        "second_url",
        "second_start",
        "second_duration"
      ],
      "properties": {
        "first_url": {
          "type": "string",
          "example": "https://www.youtube.com/watch?v=GY2fE5xr6XU"
        },
        "first_start": {
          "type": "string",
          "example": "00:01:34"
        },
        "first_duration": {
          "type": "string",
          "example": "30"
        },
        "second_url": {
          "type": "string",
          "example": "https://www.youtube.com/watch?v=hgW86REDHa0"
        },
        "second_start": {
          "type": "string",
          "example": "00:00:33"
        },
        "second_duration": {
          "type": "string",
          "example": "30"
        }
      }
    }
  },
  "externalDocs":{
    "description":"Find out more about Swagger",
    "url":"http://swagger.io"
  }
}